+++ src/commands/blacklist/add.ts
@@ -7,7 +7,7 @@ import {
     User,
 } from 'discord.js';
 
-import { CustomRole } from '../../models/enums/index.js';
+import { CustomRole } from '../../enums/role.js';
 import { EventData } from '../../models/index.js';
 import { BlacklistRepo } from '../../services/database/repos/index.js';
 import { Lang } from '../../services/index.js';
