+++ src/commands/config/require-all-trusted-roles.ts
@@ -5,7 +5,7 @@ import {
     PermissionString,
 } from 'discord.js';
 
-import { CustomRole } from '../../models/enums/index.js';
+import { CustomRole } from '../../enums/role.js';
 import { EventData } from '../../models/index.js';
 import { GuildRepo } from '../../services/database/repos/index.js';
 import { Lang } from '../../services/index.js';
